+++ packages/anvil-embed-frame/test/src/index.test.js
@@ -4,14 +4,12 @@ import AnvilEmbedFrame from '../../src/index'
 describe('AnvilEmbedFrame', function () {
   def('handleEvent', () => sinon.spy())
   def('anvilURL', 'https://app.useanvil.com')
-  def('enableDefaultStyles', true)
 
   def('render', () => shallow(
     <AnvilEmbedFrame
       iframeURL="http://localhost"
       onEvent={$.handleEvent}
       anvilURL={$.anvilURL}
-      enableDefaultStyles={$.enableDefaultStyles}
     />,
   ))